A couple lexer fixes to prevent infinite loops and overflows.
The ANY expression used to mean any byte at all, but now it's
any byte but zero, since zero is used to signal end-of-input
and it's never legal anyway (lots of places in the lexer will
check for it seperately, to give proper errors, but having it
match in ANY caused a ton of problems).

Also, since YYCURSOR can slide past YYLIMIT, it wasn't
sufficient to check for equality to see if we're at the end
of input, so adustments were made.