preprocessor: Find include files in the same directory.
So if you #include "a/b.h", and it includes "c.h", it should try to
open "a/c.h" before it tries the actual search path. It has to deal
with the initial file, which might not be in the include path, too.

Some other robustness and include improvements, too.

Fixes #2.